make fill-image tests less sensitive

+++ src/mbgl/renderer/painter.hpp
@@ -205,7 +205,7 @@ private:
     RenderPass pass = RenderPass::Opaque;
 
     int numSublayers = 3;
-    int currentLayer;
+    size_t currentLayer;
     float depthRangeSize;
     const float depthEpsilon = 1.0f / (1 << 16);
